backend/requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djoser==2.2.0
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
gunicorn==21.2.0
